+++ packages/js/src/server/integrations/uncaught_exception_monitor.ts
@@ -10,7 +10,7 @@ export default class UncaughtExceptionMonitor {
 
   constructor() {
     this.__isReporting = false
-    this.__handlerAlreadyCalled = false 
+    this.__handlerAlreadyCalled = false
     this.__listener = this.makeListener()
     this.removeAwsLambdaListener()
   }
@@ -20,17 +20,18 @@ export default class UncaughtExceptionMonitor {
   }
 
   makeListener() {
+    // noinspection UnnecessaryLocalVariableJS
     const honeybadgerUncaughtExceptionListener = (uncaughtError: Error) => {
       if (this.__isReporting || !this.__client) { return }
-    
+
       // report only the first error - prevent reporting recursive errors
       if (this.__handlerAlreadyCalled) {
         if (!this.hasOtherUncaughtExceptionListeners()) {
-          fatallyLogAndExit(uncaughtError)
+          fatallyLogAndExit(uncaughtError, 'uncaught exception')
         }
         return
       }
-    
+
       this.__isReporting = true
       this.__client.notify(uncaughtError, {
         afterNotify: (_err, _notice) => {
@@ -38,7 +39,7 @@ export default class UncaughtExceptionMonitor {
           this.__handlerAlreadyCalled = true
           this.__client.config.afterUncaught(uncaughtError)
           if (!this.hasOtherUncaughtExceptionListeners()) {
-            fatallyLogAndExit(uncaughtError)
+            fatallyLogAndExit(uncaughtError, 'uncaught exception')
           }
         }
       })
@@ -71,14 +72,14 @@ export default class UncaughtExceptionMonitor {
    * we want to report the exception to Honeybadger and
    * mimic the default behavior of NodeJs,
    * which is to exit the process with code 1
-   * 
+   *
    * Node sets up domainUncaughtExceptionClear when we use domains.
    * Since they're not set up by a user, they shouldn't affect whether we exit or not
    */
   hasOtherUncaughtExceptionListeners(): boolean {
     const allListeners = process.listeners('uncaughtException')
     const otherListeners = allListeners.filter(listener => (
-      listener.name !== 'domainUncaughtExceptionClear' 
+      listener.name !== 'domainUncaughtExceptionClear'
       && listener !== this.__listener
     ))
     return otherListeners.length > 0