+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.9.11](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/nextjs@5.9.10...@honeybadger-io/nextjs@5.9.11) (2025-03-18)
+
+**Note:** Version bump only for package @honeybadger-io/nextjs
+
+
+
+
+
 ## [5.9.10](https://github.com/honeybadger-io/honeybadger-js/compare/@honeybadger-io/nextjs@5.9.9...@honeybadger-io/nextjs@5.9.10) (2024-12-09)